+++ CHANGELOG.md
@@ -3,6 +3,7 @@ musicscan CHANGELOG
 
 ## CURRENT
 
+ - [musicscan-81](https://github.com/cjcodeproj/musicscan/issues/81) Rate limit file generation
  - [musicscan-95](https://github.com/cjcodeproj/musicscan/issues/93) Test against TinyTag 2.2.1
  - [musicscan-42](https://github.com/cjcodeproj/musicscan/issues/42) Lyric dump as XML element
  - [musicscan-91](https://github.com/cjcodeproj/musicscan/issues/91) id3scan output should list manifest file